Исправления после ревью

+++ src/ru/yandex/javacourse/kanban/manager/InMemoryHistoryManager.java
@@ -1,12 +1,14 @@
 package ru.yandex.javacourse.kanban.manager;
 
 import ru.yandex.javacourse.kanban.tasks.Task;
+
 import java.util.ArrayList;
+import java.util.List;
 
-public class InMemoryHistoryManager implements HistoryManager{
+public class InMemoryHistoryManager implements HistoryManager {
 
-    private final ArrayList<Task> historyDb;
-    private final int MAX_HISTORY_COUNT = 10;
+    private final List<Task> historyDb;
+    private static final int MAX_HISTORY_COUNT = 10;
 
     public InMemoryHistoryManager() {
         this.historyDb = new ArrayList<>();
@@ -14,6 +16,9 @@ public class InMemoryHistoryManager implements HistoryManager{
 
     @Override
     public void add(Task task) {
+        if (task == null) {
+            return;
+        }
         if (historyDb.size() == MAX_HISTORY_COUNT){
             historyDb.remove(0);
         }
@@ -21,10 +26,12 @@ public class InMemoryHistoryManager implements HistoryManager{
     }
 
     @Override
-    public ArrayList<Task> getHistory() {
-        ArrayList<Task> list = new ArrayList<>();
-        list.addAll(historyDb);
-        return list;
+    public List<Task> getHistory() {
+        return new ArrayList<>(historyDb);
     }
 
+    @Override
+    public void clearHistory() {
+        historyDb.clear();
+    }
 }
